Update actisoftR.functions and examples.r
Hi @EdgarSantos-Fernandez -- updated a couple of small things. This is looking great so far. One instance of `rbind.fill()` to replace for you noted in commenting in the code. 
Do you have what you need to start working on the summary report functions? They should be produced using the sample parameter files we put in Dropbox. A good one to work with is here: https://www.dropbox.com/s/1fvui9g5dk6kxec/ReportPeriod%20Parameter%20Sets.csv?dl=0 because it has some tricky scenarios built in. It uses this data for sample data: https://www.dropbox.com/s/1fvui9g5dk6kxec/ReportPeriod%20Parameter%20Sets.csv?dl=0
You can also look at these sample parameter files: https://www.dropbox.com/s/lait6ftjpjcoqlm/input%20parameters%20example%20period.csv?dl=0 and https://www.dropbox.com/s/skzpgemw4cvq2tz/input%20parameters%20example%20point.csv?dl=0 to be used with the sample data already used in the plots above.

Regarding the subject_ID being repeated -- it is possible for a participant to have more than one actigraphy file for a number of reasons. It is not typically during the same time period, but again it is possible. The reason why there are dates repeated in the sample data is because we were testing multiple watches on the same wrist at the same time. A typical example of 2 data files per person would be if we are monitoring someone across a pre- and post-intervention, so we have one file with their pre data and one with their post data. In this case the subject_ID would be the same but the file name would be different.