Update read_actigraph_csv.r
Added `droplevels()` to the end of `alldata2` because old factor levels were still showing up when I look at the structure of the data (i.e., "Down", "O - O", and "Up" show up when you run `levels(acti_data$interval_type)`). Please feel free to change/move if this is not the right place for it.