update
added lines 24-44. For cases in which files from AMI and Actiware are in a single folder. The function will first create the folders AMI and Actiware in the path (if they don't exist), then will copy all the AMI files into the AMI folder and Actiware files to the Actiware folder. Then the usual procedure is applied to import all the CSV files. If there is at least a file that is not of the from AMI or Actiware in the folder it will print a warning. Hope this makes sense for you. Please, let me know if you are Ok with it.