updated issues with the tz
The argument tz is used in both plot functions as (baseline) x-axis time. If the data frame contains Date/time data with tz included, the function will process internally with this time zone, but the plot will show in the specified tz argument. If datime_start and datime_end are character variables, then the plot functions will plug in the tz. If tz is missing then UTC will be used.